2sem/requirements.txt
alembic==1.15.2
amqp==5.3.1
annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1
bcrypt==4.3.0
billiard==4.2.1
celery==5.4.0
cffi==1.17.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cryptography==44.0.2
ecdsa==0.19.1
fastapi==0.115.12
gevent==24.11.1
greenlet==3.1.1
h11==0.14.0
idna==3.10
jwt==1.3.1
kombu==5.5.1
Mako==1.3.9
MarkupSafe==3.0.2
passlib==1.7.4
prompt_toolkit==3.0.50
psycopg2==2.9.10
psycopg2-binary==2.9.10
pyasn1==0.4.8
pycparser==2.22
pydantic==2.11.1
pydantic_core==2.33.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-jose==3.4.0
python-multipart==0.0.20
redis==5.2.1
rsa==4.9
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.46.1
typing-inspection==0.4.0
typing_extensions==4.13.0
tzdata==2025.1
uvicorn==0.34.0
vine==5.1.0
wcwidth==0.2.13
websocket==0.2.1
websockets==15.0.1
zope.event==5.0
zope.interface==7.2
python-jose[cryptography]>=3.3.0,<4.0.0
python-multipart>=0.0.6
pydantic-settings>=2.0.0
pydantic>=2.0.0
